pyproject.toml
[project]
name = "dictionary-mvp"
version = "0.1.0"
description = "Chinese-Russian dictionary with new features"
authors = [
    {name = "Haohanmaiyami",email = "kharitonovayan2018@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.46,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "sqladmin (>=0.22.0,<0.23.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "jinja2 (>=3.1.6,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
packages = [{ include = "dictapp", from = "src" }]
[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
